EventScores class mounted on App. Form Validation OK

+++ database/migrations/2017_12_27_145623_create_event_scores_table.php
@@ -17,7 +17,7 @@ class CreateEventScoresTable extends Migration
             $table->increments('id');
             $table->integer('team_id');
             $table->integer('event_id');
-            $table->integer('score')->nullable();
+            $table->decimal('score')->nullable();
             $table->timestamps();
         });
     }
